+++ ui/components/stimulation/StimulationStudioBlockViewEditor.vue
@@ -1,19 +1,6 @@
 <template>
   <div class="div__BlockViewEditor-background">
-    <div class="div__Tabs-panel">
-      <span
-        :id="'Basic'"
-        :class="active_tab === 'Advanced' ? 'span__Inactive-Tab-labels' : 'span__Active-Tab-label'"
-        @click="toggle_tab($event.target.id)"
-        >Basic</span
-      >
-      <span
-        :id="'Advanced'"
-        :class="active_tab === 'Basic' ? 'span__Inactive-Tab-labels' : 'span__Active-Tab-label'"
-        @click="toggle_tab($event.target.id)"
-        >Advanced</span
-      >
-    </div>
+    <div class="div__spacer" />
     <div class="div__Editor-background">
       <div class="div__setting-panel-container">
         <span :key="current_letter" class="span__protocol-letter" :style="'color:' + current_color">{{
@@ -103,7 +90,6 @@ Vue.component("BModal", BModal);
 library.add(faTrashAlt);
 
 /**
- * @vue-data {String} active_tab - Shows current selected tab
  * @vue-data {Boolean} disabled - Disables the name and time input fields
  * @vue-data {String} current_letter - Next available letter in alphabet
  * @vue-data {String} current_color -  Next available color in alphabet
@@ -117,7 +103,6 @@ library.add(faTrashAlt);
  * @vue-data {Array} protocol_list - All available protocols from Vuex
  * @vue-event {Event} update_protocols - Gets called when a change to the available protocol list occurs to update next available color/letter assignment and dropdown options
  * @vue-event {Event} handle_trash_modal - Toggle view of delete popover on trash icon
- * @vue-event {Event} toggle_tab - Toggles which tab is active
  * @vue-event {Event} handle_delete - Confirms and commits the deletion of protocol to state
  * @vue-event {Event} handle_stop_setting - Currently just assigns the new stop setting to local state
  * @vue-event {Event} handle_rest_duration - Commits the new delay input to state
@@ -134,7 +119,6 @@ export default {
   },
   data() {
     return {
-      active_tab: "Basic",
       disabled_time: false,
       current_letter: "",
       current_color: "",
@@ -219,15 +203,14 @@ export default {
       this.current_letter = letter;
       this.current_color = color;
     },
-    toggle_tab(tab) {
-      tab === "Basic" ? (this.active_tab = "Basic") : (this.active_tab = "Advanced");
-    },
     open_del_modal() {
       this.$bvModal.show("del-protocol-modal");
     },
     close_del_protocol_modal(idx) {
       this.$bvModal.hide("del-protocol-modal");
-      if (idx === 0) this.handle_protocol_editor_reset();
+      if (idx === 0) {
+        this.handle_protocol_editor_reset();
+      }
     },
     handle_stop_setting(idx) {
       const setting = this.stop_options_array[idx];
@@ -235,7 +218,9 @@ export default {
 
       this.disabled_time = idx === 1;
 
-      if (this.disabled_time) this.handle_rest_duration("0");
+      if (this.disabled_time) {
+        this.handle_rest_duration("0");
+      }
 
       this.set_stop_setting(setting.includes("Stopped"));
     },
@@ -408,14 +393,7 @@ export default {
   height: 166px;
 }
 
-.div__Tabs-panel {
-  background: rgb(17, 17, 17);
-  width: 200px;
+.div__spacer {
   height: 28px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 12px;
-  cursor: pointer;
 }
 </style>