+++ ui/components/stimulation/StimulationStudioCreateAndEdit.vue
@@ -7,9 +7,11 @@
     <div class="div__stimulationstudio-select-dropdown-container">
       <SelectDropDown
         :options_text="protocol_list"
+        :show_delete_option="show_delete_option"
         :input_width="input_width"
         :input_height="input_height"
         @selection-changed="selected_protocol_change"
+        @handle-delete="handle_delete"
       />
     </div>
     <canvas class="canvas__stimulationstudio-button-separator" />
@@ -118,9 +120,14 @@ export default {
       if (idx === 0) {
         this.set_edit_mode_off();
         this.reset_protocol_editor();
-      } else await this.edit_selected_protocol(selected_protocol);
+      } else {
+        await this.edit_selected_protocol(selected_protocol);
+      }
 
       this.$emit("handle_selection_change", selected_protocol);
+    },
+    handle_delete(item) {
+      this.$bvModal.show("del-protocol-modal");
     },
     disable_selection_btn(idx) {
       return this.disable_edits || (this.selected_protocol_idx === 0 && idx === 0);
@@ -167,6 +174,9 @@ export default {
     },
     handle_export() {
       this.handle_export_protocol();
+    },
+    show_delete_option(p) {
+      return p.letter.trim().length !== 0;
     },
   },
 };